Added support for char-level training.
